requirements.txt
flask
scikit-learn
pandas
numpy
matplotlib
joblib
fpdf2
requests
gunicorn
